+++ DefaultTypeVisitor.php
@@ -61,7 +61,7 @@ abstract class DefaultTypeVisitor implements TypeVisitor
         return $this->default($self);
     }
 
-    public function int(Type $self/** ?int $min, ?int $max */): mixed
+    public function int(Type $self, ?int $min, ?int $max): mixed
     {
         return $this->default($self);
     }
@@ -73,16 +73,6 @@ abstract class DefaultTypeVisitor implements TypeVisitor
 
     public function intMask(Type $self, Type $type): mixed
     {
-        return $this->default($self);
-    }
-
-    /**
-     * @deprecated will be removed in 0.4.0
-     */
-    public function intRange(Type $self, ?int $min, ?int $max): mixed
-    {
-        trigger_deprecation('typhoon/reflection', '0.3.3', 'Method %s() is deprecated and will be removed in 0.4.0. Use %s::int() instead.', __METHOD__, self::class);
-
         return $this->default($self);
     }