+++ TypeVisitor.php
@@ -78,7 +78,7 @@ interface TypeVisitor
      * @param Type<int> $self
      * @return TReturn
      */
-    public function int(Type $self/** ?int $min, ?int $max */): mixed;
+    public function int(Type $self, ?int $min, ?int $max): mixed;
 
     /**
      * @param non-empty-list<Type> $types
@@ -91,13 +91,6 @@ interface TypeVisitor
      * @return TReturn
      */
     public function intMask(Type $self, Type $type): mixed;
-
-    /**
-     * @deprecated will be removed in 0.4.0
-     * @param Type<int> $self
-     * @return TReturn
-     */
-    public function intRange(Type $self, ?int $min, ?int $max): mixed;
 
     /**
      * @param Type<iterable<mixed>> $self