+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Removed `TypeVisitor::value()` and `DefaultVisitor::value()`.
 - Removed `TypeVisitor::classStringLiteral()` and `DefaultVisitor::classStringLiteral()`.
 - Added `@no-named-arguments` phpDocs to `types::alias()`, `types::template()` and `types::object()`.
+- Added `?int $min` and `?int $max` parameters to `TypeVisitor` and `DefaultVisitor`.
+- Removed `TypeVisitor::intRange()` and `DefaultVisitor::intRange()` methods.
 
 ## 0.3.3